cloud_ai_layer/requirements.txt
# Cloud AI Layer Dependencies

# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data Handling and Utilities
pandas>=1.3.0
typing-extensions>=4.0.0

# External Data Integration
requests>=2.26.0

# Configuration and Environment
python-dotenv>=0.19.0

# Logging
structlog>=21.1.0

# Optional: Advanced Caching
cachetools>=5.0.0

# Optional: Performance Optimization
joblib>=1.1.0